Moved sharedpreferences implementation from VoteActivity to LoginActivity. Implemented SQLite usage in VoteActivity instead.
