+++ LearningApp/app/src/main/java/com/example/justina/learningapp/data/db/DBContract.java
@@ -1,4 +1,4 @@
-package com.example.justina.learningapp.data;
+package com.example.justina.learningapp.data.db;
 
 import android.provider.BaseColumns;
